+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Change Log
 
+## [1.1.0] - 2026-04-19
+
+### Changed
+- Added Blockquote support
+- Added stacking Bold and Italic support
+- Add horizontal line via command palette
+
 ## [1.0.0] - 2026-04-19
 
 ### Changed
 - Init project **VSC Markdown Formatter** (publisher: Studio Rude Box)
-- Reset version to 1.0.0 as the baseline